generic license information
we have to license things separately, since we're using some images,
etc, in the sample websites that we can't put under a free license and
different licenses are better suited to some of the content. to avoid
having a blank LICENSE document in the root, let's tell people to look
to individual directories.

(now we just have to remember to put a LICENSE document in each
directory!)